refactor(core): relocate write pipeline into atm_core::write, break ack<->send cycle
Create crates/atm-core/src/write as the canonical write module owning the
full write pipeline: AckRequest/AckOutcome/AckReplyDisposition/ReplyTarget,
WriteOutcome, PreparedWrite, the send_mail/write_mail entry family, the
acknowledgement admission pipeline, and received-hook dispatch planning.

- ack becomes a thin facade: types re-exported from crate::write,
  ack_mail_with_runtime delegates to crate::write::write_mail_with_runtime
- send keeps its staying helpers (persistence, recipient resolution,
  classification, hook prep) and re-exports the write API pub-use only,
  so external paths are unchanged
- dependency direction is now ack -> write, write -> send (helpers only),
  send -> write via re-exports; the sc-boundary ack<->send architectural
  cycle is cleared by identity with zero new findings
- widen pub(super) helpers consumed by write to pub(crate) at definition
  (resolve_message_body, prepare_received_hook, prepare_send_context,
  async persistence helpers, VerifiedTemplateSend)
- re-point both architecture tripwires (second-write-pipeline and AL3
  received-hook single-path) at the write module and strengthen them to
  scan send and write for prohibited detached-work constructs

AU.3 work items 2-4 (sprint doc: docs/plans/phase-au/sprint-AU3-ack-send-write-module.md)
